requirements.txt
# XFedCrop – PyTorch + XAI + API (matches project plan)
# Optional: Flower + Ray for framework-native FL — pip install "flwr[simulation]"
torch>=2.1.0
torchvision>=0.16.0
# Optional: pip install captum (falls back to PyTorch-only Grad-CAM if missing)
# captum>=0.7.0
numpy>=1.24.0
Pillow>=10.0.0
tqdm>=4.66.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
lime>=0.2.0.1
shap>=0.44.0
pytest>=7.4.0
matplotlib>=3.7.0
